+++ tomcat/src/main/java/org/apache/coyote/http11/controller/Controller.java
@@ -1,9 +1,9 @@
-package org.apache.coyote.http11.handler;
+package org.apache.coyote.http11.controller;
 
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
 
-public interface Handler {
+public interface Controller {
 
     void handle(HttpRequest httpRequest, HttpResponse httpResponse);
 }
